알림음 추가 + 코드 정리
AVFoundataion으로 카운트다운 소리 추가

코드 정리

코드 및 폴더 구조 정리
