+++ WOD/Timer/Common/View/Setting/MovementsSet.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct MovementsSet: View {
-    @EnvironmentObject var viewModel: WodViewModel
+    @EnvironmentObject var viewModel: SimpleViewModel
     @StateObject var manager = InputManager()
 
     @State private var isChange: Bool = false
