+++ WOD/Timer/Common/View/Setting/PreparationSet.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct PreparationSet: View {
-    @EnvironmentObject var viewModel: WodViewModel
+    @EnvironmentObject var viewModel: SimpleViewModel
     @StateObject var manager = InputManager()
     @State private var isChange: Bool = false
