+++ WOD/Timer/ContentView.swift
@@ -8,8 +8,20 @@
 import SwiftUI
 
 struct ContentView: View {
+    @State var selectedTab = 0
+    
     var body: some View {
-        MainView()
+        TabView {
+            SimpleView()
+                .tabItem {
+                    Image(systemName: "figure.highintensity.intervaltraining")
+                }
+            
+            StopWatchView()
+                .tabItem {
+                    Image(systemName: "timer")
+                }
+        }
     }
 }
