+++ WOD/Timer/Simple/InputModel.swift
