+++ WOD/Timer/Simple/Model/Enum.swift
